jwt_proxy/requirements.txt
requests
aiohttp
pyjwt
python-dotenv
pytest-aiohttp
redis
pylint
coverage
cchardet
aiodns
asyncio